<fix> Add confirm delete dialog for delete employee button
This commit adds confirm dialog when the delete button is clicked to confirm the action before implementing the delete employee function.

- Added confirmation dialog using window.comfirm
- Updated view after deleting employee

+++ Employee MS/src/Components/Dashboard/Employee.jsx
@@ -31,12 +31,26 @@ const Employee = () => {
     }
 
     const handleDeleteEmployee = async (_id) => {
-        let results = await adminService.deleteEmployeeService(_id);
-        if (results && results.status === 204) {
-            window.location.reload();
+        const confirmDelete = window.confirm("Are you sure you want to delete this employee?");
+
+        if (!confirmDelete) {
+            return;
         }
-        else {
-            //error
+
+        try {
+            let results = await adminService.deleteEmployeeService(_id);
+
+            if (results && results.status === 204) {
+                const updatedEmployees = employee.filter(employee => employee._id !== _id);
+                setEmployee(updatedEmployees);
+
+                toast.success("Employee deleted successful");
+            } else {
+                toast.error("Failed to delete employee");
+            }
+        } catch (error) {
+            console.error("Error deleting employee:", error);
+            toast.error("An error occurred while deleting employee");
         }
     }
